* Withdrawal controller test & impl * Withdrawal service test & impl * Minor refactor

+++ src/main/java/com/wezaam/withdrawal/service/WithdrawalProcessingService.java
@@ -4,10 +4,12 @@ import com.wezaam.withdrawal.exception.TransactionException;
 import com.wezaam.withdrawal.model.PaymentMethod;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class WithdrawalProcessingService {
 
-    public Long sendToProcessing(Double amount, PaymentMethod paymentMethod) throws TransactionException {
+    public Long sendToProcessing(BigDecimal amount, PaymentMethod paymentMethod) throws TransactionException {
         // call a payment provider
         // in case a transaction can be process
         // it generates a transactionId and process the transaction async
